+++ src/potions.c
@@ -1,6 +1,6 @@
 /*
  * potions.c
- * Copyright (C) 2009-2018 Joachim de Groot <jdegroot@web.de>
+ * Copyright (C) 2009-2021 Joachim de Groot <jdegroot@web.de>
  *
  * NLarn is free software: you can redistribute it and/or modify it
  * under the terms of the GNU General Public License as published by the
@@ -273,8 +273,12 @@ item_usage_result potion_quaff(struct player *p, item *potion)
         case PO_MAX_HP:
             result.identified = potion_with_effect(p, potion);
 
-            // Potions of instant healing can cure poisoning and sickness,
-            // the latter is performed in player_effect_add().
+            // Potions of instant healing can cure poisoning and sickness
+            if ((e = player_effect_get(nlarn->p, ET_SICKNESS)))
+            {
+                player_effect_del(nlarn->p, e);
+                result.identified = TRUE;
+            }
             if ((e = player_effect_get(nlarn->p, ET_POISON)))
             {
                 player_effect_del(nlarn->p, e);